NSDEBase.jl

This is the documentation of NSDEBase.jl, a Julia package containing common types and utility functions shared by NSDE.jl and its subpackages.

Manual

NSDEBase.jl

Manual

API

Abstract types

Composite types

ODEs

Utilities

Index

API

All exported types and functions are considered part of the public API, and thus documented in this manual.

Abstract types

Missing docstring.

Missing docstring for AbstractNSDEType. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractNSDEProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractNSDESolver. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractNSDESolution. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractInitialValueProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractInitialValueSolver. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractInitialValueSolution. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractRightHandSide. Check Documenter's build log for details.

Composite types

NSDEBase.InitialValueProblem — Type

InitialValueProblem <: AbstractInitialValueProblem

A composite type for an initial-value problem.

Constructors

InitialValueProblem(rhs, u0, tspan)
InitialValueProblem(rhs, u0, t0, tN)
IVP(args...; kwargs...)

Arguments

rhs :: Union{AbstractRightHandSide, Function, AbstractMatrix{ℂ}, ℂ} where ℂ<:Number

u0 :: Union{AbstractVector{ℂ}, ℂ} where ℂ<:Number

tspan :: Tuple{ℝ, ℝ} where ℝ<:Real

Functions

subproblemof : creates a subproblem.

NSDEBase.subproblemof — Function

subproblemof(problem, u0, tspan) :: InitialValueProblem
subproblemof(problem, u0, t0, tN) :: InitialValueProblem

returns a subproblem of problem, i.e. a copy with same rhs but different u0 and tspan.

Missing docstring.

Missing docstring for RightHandSide. Check Documenter's build log for details.

NSDEBase.SplitRightHandSide — Type

SplitRightHandSide <: AbstractRightHandSide

A composite type for the right-hand side of an InitialValueProblem in the form $f_s(u, t) + f_{ns}(u, t)$.

Constructors

SplitRightHandSide(stiff, nonstiff)
SRHS(args...; kwargs...)

Arguments

stiff :: Union{NonlinearRightHandSide, LinearRightHandSide, Function, AbstractMatrix{ℂ}, ℂ} where ℂ<:Number

nonstiff :: Union{NonlinearRightHandSide, Function}

ODEs

NSDEBase.Dahlquist — Function

Dahlquist(u0=0.5, tspan=(0.0, 1.0); λ=1.0) :: InitialValueProblem
Dahlquist(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Dahlquist equation.

NSDEBase.Logistic — Function

Logistic(u0=0.5, tspan=(0.0, 1.0); λ=1.0) :: InitialValueProblem
Logistic(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Logistic equation.

NSDEBase.SimplePendulum — Function

SimplePendulum(u0=[π/4, 0.0], tspan=(0.0, 2π/3 * √9.81)) :: InitialValueProblem
SimplePendulum(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the simple pendulum problem.

Missing docstring.

Missing docstring for VanderPol. Check Documenter's build log for details.

NSDEBase.Rössler — Function

Rössler(u0=[2.0, 0.0, 0.0], tspan=(0.0, 1.0); α=0.2, β=0.2, γ=5.7) :: InitialValueProblem
Rössler(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Rössler equations.

NSDEBase.Lorenz — Function

Lorenz(u0=[2.0, 3.0, -14.0], tspan=(0.0, 1.0); σ=10.0, β=8/3, ρ=28.0) :: InitialValueProblem
Lorenz(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Lorenz equations.

NSDEBase.Lorenz96 — Function

Lorenz96(u0=[ones(39); 1.01], tspan=(0.0, 1.0); F=8.0) :: InitialValueProblem
Lorenz96(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Lorenz-96 equations.

Utilities

Base.show — Function

show(io::IO, object::AbstractShowable)

prints the full description of an object and its contents to a stream io.

Base.summary — Function

summary(io::IO, object::AbstractShowable)

prints the short description of an object to a stream io.

Index

NSDEBase.InitialValueProblem

NSDEBase.SplitRightHandSide

Base.show

Base.summary

NSDEBase.Dahlquist

NSDEBase.Logistic

NSDEBase.Lorenz

NSDEBase.Lorenz96

NSDEBase.Rössler

NSDEBase.SimplePendulum

NSDEBase.subproblemof