NSDEBase.jl

This is the documentation of NSDEBase.jl, a Julia package containing common types and utility functions shared by NSDE.jl and its subpackages.

Manual

NSDEBase.jl

Manual

API

Abstract types

Composite types

ODEs

Utilities

Index

API

All exported types and functions are considered part of the public API, and thus documented in this manual.

Abstract types

Missing docstring.

Missing docstring for AbstractObject. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractSolver. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractSolution. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractParameters. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractCache. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractInitialValueProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractInitialValueSolver. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractInitialValueSolution. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractInitialValueParameters. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractInitialValueCache. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractRightHandSide. Check Documenter's build log for details.

Composite types

NSDEBase.LinearRightHandSide — Type

LinearRightHandSide <: AbstractRightHandSide

A composite type for the right-hand side of an InitialValueProblem in the form $Lu + g(t)$.

Constructors

LinearRightHandSide(L, g, g!)
LinearRightHandSide(L[, g!_or_g])
LRHS(args...; kwargs...)

Arguments

L :: Union{Number,AbstractMatrix{<:Number}} : $L$.

g :: Union{Function,Nothing} : $g$.

g! :: Union{Function,Nothing} : $g$ (in-place).

NSDEBase.NonlinearRightHandSide — Type

NonlinearRightHandSide <: AbstractRightHandSide

A composite type for the right-hand side of an InitialValueProblem in the form $f(u, t)$.

Constructors

NonlinearRightHandSide(f, f!, Df, Df!)
NonlinearRightHandSide(f!_or_f, is_complex=false)
RightHandSide(args...; kwargs...)
RHS(args...; kwargs...)

Arguments

f :: Function : $f$. f! :: Function : $f$ (in-place). Df :: Function : Jacobian of $f$. Df! :: Function : Jacobian of $f$ (in-place).

NSDEBase.SplitRightHandSide — Type

SplitRightHandSide <: AbstractRightHandSide

A composite type for the right-hand side of an InitialValueProblem in the form $f_s(u, t) + f_{ns}(u, t)$.

Constructors

SplitRightHandSide(stiff, nonstiff)
SRHS(args...; kwargs...)

Arguments

stiff :: Union{NonlinearRightHandSide, LinearRightHandSide, Function, AbstractMatrix{ℂ}, ℂ} where ℂ<:Number : $f_s$.

nonstiff :: Union{NonlinearRightHandSide, Function} : $f_{ns}$.

NSDEBase.InitialValueProblem — Type

InitialValueProblem <: AbstractInitialValueProblem

A composite type for an initial-value problem.

Constructors

InitialValueProblem(rhs, u0, tspan)
InitialValueProblem(rhs, u0, t0, tN)
IVP(args...; kwargs...)

Arguments

rhs :: Union{AbstractRightHandSide, Function, AbstractMatrix{ℂ}, ℂ} where ℂ<:Number : right-hand side function.

u0 :: Union{AbstractVector{ℂ}, ℂ} where ℂ<:Number : initial condition.

tspan :: Tuple{ℝ, ℝ} where ℝ<:Real : end limits of time domain.

Functions

subproblemof : creates a subproblem.

NSDEBase.subproblemof — Function

subproblemof(problem, u0, tspan) :: InitialValueProblem
subproblemof(problem, u0, t0, tN) :: InitialValueProblem

returns a subproblem of problem with the same rhs but different u0 and tspan.

ODEs

NSDEBase.Dahlquist — Function

Dahlquist(u0=0.5, tspan=(0.0, 1.0); λ=1.0) :: InitialValueProblem
Dahlquist(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Dahlquist equation.

NSDEBase.Logistic — Function

Logistic(u0=0.5, tspan=(0.0, 1.0); λ=1.0) :: InitialValueProblem
Logistic(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Logistic equation.

NSDEBase.SimplePendulum — Function

SimplePendulum(u0=[π/4, 0.0], tspan=(0.0, 2π/3 * √9.81)) :: InitialValueProblem
SimplePendulum(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the simple pendulum problem.

NSDEBase.DoublePendulum — Function

DoublePendulum(u0=[π/4, π/4, 0.0, 0.0], tspan=(0.0, 1.0); μ=1.0, λ=1.0) :: InitialValueProblem
DoublePendulum(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the double pendulum problem.

NSDEBase.VanDerPol — Function

VanDerPol(u0=[1.0, 0.0], tspan=(0.0, 1.0); μ=1.0) :: InitialValueProblem
VanDerPol(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Van der Pol equation (in first-order form).

NSDEBase.Rössler — Function

Rössler(u0=[2.0, 0.0, 0.0], tspan=(0.0, 1.0); α=0.2, β=0.2, γ=5.7) :: InitialValueProblem
Rössler(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Rössler equations.

NSDEBase.Lorenz — Function

Lorenz(u0=[2.0, 3.0, -14.0], tspan=(0.0, 1.0); σ=10.0, β=8/3, ρ=28.0) :: InitialValueProblem
Lorenz(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Lorenz equations.

NSDEBase.Lorenz96 — Function

Lorenz96(u0=[ones(39); 1.01], tspan=(0.0, 1.0); F=8.0) :: InitialValueProblem
Lorenz96(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Lorenz-96 equations.

Utilities

Base.show — Function

show(io::IO, object::AbstractObject)

prints the full description of an object and its contents to a stream io.

Base.summary — Function

summary(io::IO, object::AbstractObject)

prints the short description of an object to a stream io.

Index

NSDEBase.InitialValueProblem

NSDEBase.LinearRightHandSide

NSDEBase.NonlinearRightHandSide

NSDEBase.SplitRightHandSide

Base.show

Base.summary

NSDEBase.Dahlquist

NSDEBase.DoublePendulum

NSDEBase.Logistic

NSDEBase.Lorenz

NSDEBase.Lorenz96

NSDEBase.Rössler

NSDEBase.SimplePendulum

NSDEBase.VanDerPol

NSDEBase.subproblemof