NSDEBase.jl

This is the documentation of NSDEBase.jl, a Julia package containing shared types and functions of NSDE.jl and its sub-packages.

Manual

NSDEBase.jl

Manual

API

Types

ODEs

Utilities

Index

API

All exported types and functions are considered part of the public API and thus documented in this manual.

Types

NSDEBase.NSDEProblem — Type

An abstract type for the problems from NSDE.jl.

NSDEBase.NSDESolver — Type

An abstract type for the solvers from NSDE.jl.

NSDEBase.NSDESolution — Type

An abstract type for the solutions from NSDE.jl.

NSDEBase.RightHandSideFunction — Type

RightHandSideFunction

A composite type for the right-hand side of an InitialValueProblem.

Constructors

RightHandSideFunction(f, f!, Df, Df!)
RightHandSideFunction(f!_or_f)
IVP(args...; kwargs...)

Arguments

f :: Function : right-hand side derivative.

f! :: Function : right-hand side derivative (in-place).

Df :: Function : jacobian of right-hand side derivative.

Df! :: Function : jacobian of right-hand side derivative (in-place).

f!_or_f :: Function : function (in-place or not) from which all other fields are automatically constructed.

Functions

show : shows name and contents.

summary : shows name.

NSDEBase.InitialValueProblem — Type

InitialValueProblem <: NSDEProblem

A composite type for initial value problems.

Constructors

InitialValueProblem(rhs, u0, tspan)
InitialValueProblem(rhs, u0, t0, tN)
IVP(args...; kwargs...)

Arguments

rhs :: Union{Function, RightHandSideFunction} : right-hand side derivative.

u0 :: Union{Number, AbstractVector{Number}} : initial condition.

tspan :: Tuple{Real, Real} : time domain.

t0 :: Real : initial time.

tN :: Real : final time.

Functions

show : shows name and contents.

summary : shows name.

NSDEBase.InitialValueSolver — Type

An abstract type for the solvers of InitialValueProblems.

NSDEBase.InitialValueSolution — Type

An abstract type for the solution of an InitialValueProblem.

ODEs

NSDEBase.Dahlquist — Function

Dahlquist(u0 = 1.0, tspan = (0.0, 1.0); λ = 1.0) :: InitialValueProblem

returns an InitialValueProblem for the Dahlquist equation.

NSDEBase.Logistic — Function

Logistic(u0 = 0.5, tspan = (0.0, 5.0); λ = 1.0) :: InitialValueProblem

returns an InitialValueProblem for the Logistic equation.

NSDEBase.Riccati — Function

Riccati(u0 = 0.0, tspan = (0.0, 5.0)) :: InitialValueProblem

returns an InitialValueProblem for the Riccati equation.

NSDEBase.SimplePendulum — Function

SimplePendulum(u0 = [0.0, π/2], tspan = (0.0, 2π); g = 1.0, L = 1.0) :: InitialValueProblem

returns an InitialValueProblem for the simple pendulum problem.

NSDEBase.VanderPol — Function

VanderPol(u0 = [1.0; 0.0], tspan = (0.0, 5.0); μ = 1.0) :: InitialValueProblem

returns an InitialValueProblem for the Van der Pol equation (in first-order form).

NSDEBase.Rössler — Function

Rössler(u0 = [2.0, 0.0, 0.0], tspan = (0.0, 10.0); a = 0.2, b = 0.2, c = 5.7) :: InitialValueProblem

returns an InitialValueProblem for the Rössler equations.

NSDEBase.Lorenz — Function

Lorenz(u0 = [2.0, 3.0, -14.0], tspan = (0.0, 10.0); σ = 10.0, β = 8/3, ρ = 28.0) :: InitialValueProblem

returns an InitialValueProblem for the Lorenz equations.

NSDEBase.Lorenz96 — Function

Lorenz96(u0, tspan; F = 8) :: InitialValueProblem

returns an InitialValueProblem for the Rössler equations.

Utilities

Base.show — Function

show(io::IO, rhs::RightHandSideFunction)

prints a full description of rhs and its contents to a stream io.

show(io::IO, problem::InitialValueProblem)

prints a full description of problem and its contents to a stream io.

Base.summary — Function

summary(io::IO, rhs::RightHandSideFunction)

prints a brief description of rhs to a stream io.

summary(io::IO, problem::InitialValueProblem)

prints a brief description of problem to a stream io.

Index

NSDEBase.InitialValueProblem

NSDEBase.InitialValueSolution

NSDEBase.InitialValueSolver

NSDEBase.NSDEProblem

NSDEBase.NSDESolution

NSDEBase.NSDESolver

NSDEBase.RightHandSideFunction

Base.show

Base.summary

NSDEBase.Dahlquist

NSDEBase.Logistic

NSDEBase.Lorenz

NSDEBase.Lorenz96

NSDEBase.Riccati

NSDEBase.Rössler

NSDEBase.SimplePendulum

NSDEBase.VanderPol