NSDEBase.jl

This is the documentation of NSDEBase.jl, a Julia package containing shared types and functions of NSDE.jl and its sub-packages.

Manual

NSDEBase.jl

Manual

API

Abstract types

Composite types

ODEs

Utilities

Index

API

All exported types and functions are considered part of the public API and thus documented in this manual.

Abstract types

NSDEBase.AbstractInitialValueProblem — Type

An abstract type for initial value problems.

NSDEBase.AbstractInitialValueSolution — Type

An abstract type for the solution of an AbstractInitialValueProblem obtained with an AbstractInitialValueSolver.

NSDEBase.AbstractInitialValueSolver — Type

An abstract type for the solvers of AbstractInitialValueProblems.

NSDEBase.AbstractNSDEObject — Type

An abstract type for any object from NSDE.jl.

NSDEBase.AbstractNSDEProblem — Type

An abstract type for the problems from NSDE.jl.

NSDEBase.AbstractNSDESolution — Type

An abstract type for the solutions from NSDE.jl.

NSDEBase.AbstractNSDESolver — Type

An abstract type for the solvers from NSDE.jl.

NSDEBase.AbstractRightHandSideFunction — Type

An abstract type for the right-hand side of an AbstractInitialValueProblem.

Composite types

NSDEBase.InitialValueProblem — Type

InitialValueProblem <: AbstractInitialValueProblem

A composite type for an AbstractInitialValueProblem.

Constructors

InitialValueProblem(rhs, u0, tspan)
InitialValueProblem(rhs, u0, t0, tN)
IVP(args...; kwargs...)

Arguments

rhs::Union{Function, AbstractRightHandSideFunction} : right-hand side derivative.

u0::Union{Number, AbstractVector{Number}} : initial condition.

tspan::Tuple{Real, Real} : time domain.

t0::Real : initial time.

tN::Real : final time.

Functions

show : shows name and contents.

summary : shows name.

NSDEBase.IVP — Method

InitialValueProblem <: AbstractInitialValueProblem

A composite type for an AbstractInitialValueProblem.

Constructors

InitialValueProblem(rhs, u0, tspan)
InitialValueProblem(rhs, u0, t0, tN)
IVP(args...; kwargs...)

Arguments

rhs::Union{Function, AbstractRightHandSideFunction} : right-hand side derivative.

u0::Union{Number, AbstractVector{Number}} : initial condition.

tspan::Tuple{Real, Real} : time domain.

t0::Real : initial time.

tN::Real : final time.

Functions

show : shows name and contents.

summary : shows name.

NSDEBase.RightHandSideFunction — Type

RightHandSideFunction <: AbstractRightHandSideFunction

A composite type for the right-hand side of an AbstractInitialValueProblem.

Constructors

RightHandSideFunction(f, f!, Df, Df!)
RightHandSideFunction(f!_or_f)
RHS(args...; kwargs...)

Arguments

f::Function : right-hand side derivative.

f!::Function : right-hand side derivative (in-place).

Df::Function : jacobian of right-hand side derivative.

Df!::Function : jacobian of right-hand side derivative (in-place).

f!_or_f::Function : function (in-place or not) from which all other fields will be constructed.

Functions

show : shows name and contents.

summary : shows name.

NSDEBase.SplitRightHandSideFunction — Type

SplitRightHandSideFunction <: AbstractRightHandSideFunction

A composite type for the split right-hand side of an AbstractInitialValueProblem.

Constructors

SplitRightHandSideFunction(L, rhs)
SRHS(args...; kwargs...)

Arguments

L::Union{Number, AbstractMatrix} : linear part of right-hand side derivative.

rhs::RightHandSideFunction : nonlinear part of right-hand side derivative.

Functions

show : shows name and contents.

summary : shows name.

NSDEBase.RHS — Method

RightHandSideFunction <: AbstractRightHandSideFunction

A composite type for the right-hand side of an AbstractInitialValueProblem.

Constructors

RightHandSideFunction(f, f!, Df, Df!)
RightHandSideFunction(f!_or_f)
RHS(args...; kwargs...)

Arguments

f::Function : right-hand side derivative.

f!::Function : right-hand side derivative (in-place).

Df::Function : jacobian of right-hand side derivative.

Df!::Function : jacobian of right-hand side derivative (in-place).

f!_or_f::Function : function (in-place or not) from which all other fields will be constructed.

Functions

show : shows name and contents.

summary : shows name.

NSDEBase.SRHS — Method

SplitRightHandSideFunction <: AbstractRightHandSideFunction

A composite type for the split right-hand side of an AbstractInitialValueProblem.

Constructors

SplitRightHandSideFunction(L, rhs)
SRHS(args...; kwargs...)

Arguments

L::Union{Number, AbstractMatrix} : linear part of right-hand side derivative.

rhs::RightHandSideFunction : nonlinear part of right-hand side derivative.

Functions

show : shows name and contents.

summary : shows name.

ODEs

NSDEBase.Dahlquist — Function

Dahlquist(u0=0.5, tspan=(0.0, 1.0); λ=1.0) :: InitialValueProblem
Dahlquist(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Dahlquist equation.

NSDEBase.Logistic — Function

Logistic(u0=0.5, tspan=(0.0, 1.0); λ=1.0) :: InitialValueProblem
Logistic(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Logistic equation.

NSDEBase.Lorenz — Function

Lorenz(u0=[2.0, 3.0, -14.0], tspan=(0.0, 1.0); σ=10.0, β=8/3, ρ=28.0) :: InitialValueProblem
Lorenz(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Lorenz equations.

NSDEBase.Lorenz96 — Function

Lorenz96(u0=[1.01; ones(40)], tspan=(0.0, 1.0); F=8) :: InitialValueProblem
Lorenz96(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Lorenz-96 equations.

NSDEBase.Rössler — Function

Rössler(u0=[2.0, 0.0, 0.0], tspan=(0.0, 1.0); a=0.2, b=0.2, c=5.7) :: InitialValueProblem
Rössler(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Rössler equations.

NSDEBase.SimplePendulum — Function

SimplePendulum(u0=[0.0, π/2], tspan=(0.0, 2π); g=1.0, L=1.0) :: InitialValueProblem
SimplePendulum(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the simple pendulum problem.

NSDEBase.VanderPol — Function

VanderPol(u0=[1.0; 0.0], tspan=(0.0, 1.0); μ=1.0) :: InitialValueProblem
VanderPol(u0, t0, tN; kwargs...) :: InitialValueProblem

returns an InitialValueProblem for the Van der Pol equation (in first-order form).

Utilities

Base.show — Method

show(io::IO, object::AbstractNSDEObject)

prints a description of an AbstractNSDEObject and all its contents to a stream io.

Base.summary — Method

summary(io::IO, problem::AbstractNSDEObject)

prints a brief description of an AbstractNSDEObject and all its contents to a stream io.

Index

NSDEBase.AbstractInitialValueProblem

NSDEBase.AbstractInitialValueSolution

NSDEBase.AbstractInitialValueSolver

NSDEBase.AbstractNSDEObject

NSDEBase.AbstractNSDEProblem

NSDEBase.AbstractNSDESolution

NSDEBase.AbstractNSDESolver

NSDEBase.AbstractRightHandSideFunction

NSDEBase.InitialValueProblem

NSDEBase.RightHandSideFunction

NSDEBase.SplitRightHandSideFunction

Base.show

Base.summary

NSDEBase.Dahlquist

NSDEBase.IVP

NSDEBase.Logistic

NSDEBase.Lorenz

NSDEBase.Lorenz96

NSDEBase.RHS

NSDEBase.Rössler

NSDEBase.SRHS

NSDEBase.SimplePendulum

NSDEBase.VanderPol